third_party/blink/web_tests/external/wpt/css/css-flexbox/intrinsic-size/row-wrap-001-expected.txt

This is a testharness.js-based test.
[FAIL] .floating-flexbox 1
  assert_equals: \n<div class="floating-flexbox" data-expected-width="100">\n      <!-- contribution: 100 -->\n      <div style="flex: 1 1 200px; width:50px;">\n        <div></div>\n      </div>\n      <!-- contribution: 50 -->\n      <div style="flex: 1 1 400px; width:50px;">\n      </div>\n    </div>\nwidth expected 100 but got 50
[FAIL] .floating-flexbox 2
  assert_equals: \n<div class="floating-flexbox" data-expected-width="100">\n      <!-- contribution: 100px\n      starts as max(specified=200, min-content=100) = 200\n      clamped down to 50 because doesn't grow\n      clamped up to 100 because auto min size is 100\n      -->\n      <div style="flex: 0 0 50px; width: 200px;">\n        <div></div>\n      </div>\n      <!-- contribution: 50px\n      starts as max(specified=50, min-content=100) = 100\n      clamped down to 50 because doesn't grow\n      auto min size is 50, so has no effect\n      -->\n      <div style="flex: 0 0 50px; width: 50px;">\n        <div></div>\n      </div>\n    </div>\nwidth expected 100 but got 200
[FAIL] .floating-flexbox 4
  assert_equals: \n<div class="floating-flexbox" data-expected-width="450">\n      <div style="flex: 1 0 50px; width: 200px;">\n        <div></div>\n      </div>\n      <div style="flex: 2 0 150px; width: 50px; margin-left: 300px;">\n        <div></div>\n      </div>\n    </div>\nwidth expected 450 but got 350
[FAIL] .floating-flexbox 5
  assert_equals: \n<div class="floating-flexbox" data-expected-width="450" style="row-gap: 24px; column-gap: 17px;">\n      <div style="flex: 1 0 50px; width: 200px;">\n        <div></div>\n      </div>\n      <div style="flex: 2 0 150px; width: 50px; margin-left: 300px;">\n        <div></div>\n      </div>\n    </div>\nwidth expected 450 but got 350
Harness: the test ran to completion.
